requirements.txt
fastapi==0.111.0
starlette==0.37.2
anyio>=4.4,<5
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.16.5
pydantic==2.5.0
pydantic-settings==2.0.3
langdetect==1.0.9
python-dotenv==1.0.0
bcrypt==4.1.1
pyjwt==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
beautifulsoup4==4.12.3
defusedxml==0.7.1
openai>=1.70.0
anthropic>=0.40.0
langfuse>=2.60.2,<3
pypdf>=4.0.0
python-docx>=1.0.0
PyYAML==6.0.1
python-multipart>=0.0.7
email-validator==2.1.0
pytest-cov==4.1.0
slowapi==0.1.9
cryptography==42.0.5
pgvector>=0.2.0
rank-bm25>=0.2.2
ruff>=0.3.0
posthog>=3.5.0
sentry-sdk[fastapi]>=2.0
chat9 @ git+https://github.com/tabularasa31/chat9-sdks.git#subdirectory=python
